--- a/.vscode/launch.json
+++ b/.vscode/launch.json
@@ -123,7 +123,8 @@
       "name": "debug stencil compiler (default config)",
       "args": [
         "${workspaceFolder}/bin/stencil",
-        "build"
+        "build",
+        "--max-workers=1"
       ],
       "protocol": "inspector"
     },