--- a/scripts/bundles/plugins/sizzle-plugin.ts
+++ b/scripts/bundles/plugins/sizzle-plugin.ts
@@ -3,16 +3,31 @@ import type { BuildOptions } from '../../utils/options';
 import { join } from 'path';
 import type { Plugin } from 'rollup';
 
+/**
+ * Bundles sizzle, a CSS selector engine, into the Stencil compiler
+ * @param opts the options being used during a build of the Stencil compiler
+ * @returns the plugin that inlines sizzle
+ */
 export function sizzlePlugin(opts: BuildOptions): Plugin {
   return {
     name: 'sizzlePlugin',
-    resolveId(id) {
+    /**
+     * A rollup build hook for resolving sizzle [Source](https://rollupjs.org/guide/en/#resolveid)
+     * @param id the importee exactly as it is written in the import statement
+     * @returns a string that resolves an import to some id
+     */
+    resolveId(id: string): string | null {
       if (id === 'sizzle') {
         return id;
       }
       return null;
     },
-    async load(id) {
+    /**
+     * A rollup build hook for loading sizzle. [Source](https://rollupjs.org/guide/en/#load)
+     * @param id the path of the module to load
+     * @returns parse5, pre-bundled
+     */
+    async load(id: string): Promise<string> {
       if (id !== 'sizzle') {
         return null;
       }
@@ -24,7 +39,13 @@ export function sizzlePlugin(opts: BuildOptions): Plugin {
   };
 }
 
-function getSizzleBundle(opts: BuildOptions, content: string) {
+/**
+ * Creates a sizzle bundle to inline
+ * @param opts the options being used during a build of the Stencil compiler
+ * @param content the sizzle source contents
+ * @returns a modified version of sizzle, wrapped in an immediately invoked function expression (IIFE)
+ */
+function getSizzleBundle(opts: BuildOptions, content: string): string {
   return `// Sizzle ${opts.sizzleVersion}
 export default (function() {
 const window = {