… scripts/bundles/plugins/module-debug-plugin.ts