<!-- Please refer to our contributing documentation for any questions on submitting a pull request, or let us know here if you need any help: https://github.com/ionic-team/stencil/blob/main/.github/CONTRIBUTING.md -->

## Pull request checklist

Please check if your PR fulfills the following requirements:
- [ ] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been reviewed and added / updated if needed (for bug fixes / features)
- [x] Build (`npm run build`) was run locally and any changes were pushed
- [x] Unit tests (`npm test`) were run locally and passed
- [x] E2E Tests (`npm run test.karma.prod`) were run locally and passed
- [x] Prettier (`npm run prettier`) was run locally and passed

## Pull request type

<!-- Please do not submit updates to dependencies unless it fixes an issue. --> 

<!-- Please try to limit your pull request to one type, submit multiple pull requests if needed. --> 

Please check the type of change your PR introduces:
- [ ] Bugfix
- [ ] Feature
- [ ] Refactoring (no functional changes, no api changes)
- [x] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):


## What is the current behavior?
<!-- Please describe the current behavior that you are modifying, or link to a relevant issue. -->

Sourcemaps are not generated for the `compiler/` in Stencil, making it difficult to work with as a developer on the project.

GitHub Issue Number: N/A


## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by this PR. -->

This PR is split into three commits.

### 1. chore(build): add sourcemap support to compiler/ 

this commit adds the minimal amount of code necessary to begin to
support sourcemaps in the compiler directory. these sourcemaps are not
used by consumers of the stencil core compiler/package, but are used by
individuals working on the compiler itself. this commit enables setting
breakpoints, inspecting variables, etc. for the compiler directory
alone.

this commit updates (and renames) the 'hackReplace' rollup transform
hook to output sourcemaps for this transformation step. this is
necessary as the hook is actively changing compiler code, which means
that a new intermediate sourcemap must be generated for rollup to be
able to generate a final sourcemap.

it also enables sourcemaps for the rollupCommonJs plugin for the same
reason.

### 2. remove moduleDebugPlugin

this commit removes `moduleDebugPlugin`. this plugin (was) responsible
for generating a comment before every bundled module like so:
```
// MODULE: <PATH_TO_MODULE>
<module_code>
...
```

`moduleDebugPlugin` is removed for two reasons:
1. sourcemaps could not be properly generated for them (they were always off by one line, which we can best estimate was due to the comment)
2. such comments are largely unnecessary with sourcemaps

the latter point makes the assumption that folks will largely be
debugging in a browser, which is a fair critique that these comments are
nice to have in a browser setting. LMK if you feel strongly here.

### 3.  add documentation to compiler related plugins 

does what it says on the tin


## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact and migration path for existing applications below. -->

## Testing

<!-- Please describe the steps you took to test the changes in this PR. -->

### Timing

I ran `npm ci`, then `npm run clean && time npm run build` multiple times on both this branch and `main`:

Main at 7085aee6f5fb36f6f3ceb35ca48c297873e59b76:
```
npm run build  79.49s user 3.61s system 157% cpu 52.665 total
npm run build  75.60s user 3.72s system 149% cpu 53.046 total
npm run build  74.94s user 3.24s system 156% cpu 49.890 total
npm run build  71.91s user 3.35s system 152% cpu 49.361 total
npm run build  73.27s user 3.40s system 152% cpu 50.303 total
~ 75.24s user avg
```

This Branch:
```
npm run build  77.90s user 4.84s system 153% cpu 54.018 total
npm run build  79.16s user 4.20s system 151% cpu 55.113 total
npm run build  78.81s user 4.44s system 155% cpu 53.531 total
npm run build  79.99s user 4.77s system 153% cpu 55.317 total
npm run build  77.95s user 4.42s system 151% cpu 54.499 total
~ 78.76s user avg
```
Which results in a few seconds average cold build time.

For a 'warm' build (where `npm run clean` was not run prior to build), I saw
- ~55.78s user avg time on Main at 7085aee6f5fb36f6f3ceb35ca48c297873e59b76
- ~60.63s user avg time on this branch
 
### Size

When I run `npm ci && npm run build && npm pack`, the size of the generated tarball is:

On Main at 7085aee6f5fb36f6f3ceb35ca48c297873e59b76:
- package size:  3.7 MB                                  
- unpacked size: 15.5 MB   

On this Branch

- package size:  3.7 MB                                  
- unpacked size: 15.5 MB   

No change in size, which is expected since [sourcemaps are excluded from the bundle](https://github.com/ionic-team/stencil/pull/3108/files#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519R12)
## Other information



<!-- Any other information that is important to this PR such as screenshots of how the component looks before and after the change. -->

1. Dependency Sourcemaps 

Stencil bundles many of its dependencies, including TypeScript, rollup plugins, etc. This PR does not generate sourcemaps for those dependencies in an effort to balance build time and usefulness. If necessary, my thought is we can always add these later.