@splitinfinities 2021-11-16
Hey this is awesome! I cannot comment on the code itself, it builds and that's great. I'd love to have a learning series on how best to take advantage of these changes in day-to-day (instead of doing console.log, which is what I did often for Telemetry...)